+++ tests/Feature/Connector/Requests/AppointmentTests.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Clinect\NextGen\Tests\Feature\Resources\Requests\Appointments;
+namespace Clinect\NextGen\Tests\Feature\Connector\Requests;
 
 use Clinect\NextGen\NextGen;
 use Orchestra\Testbench\TestCase;
 use Clinect\NextGen\Tests\Stubs\Appointment as AppointmentStub;
 
-class HealthHistoryFormTests extends TestCase
+class AppointmentTests extends TestCase
 {
     use AppointmentStub;
 
-    public function testCanSeeAllAppointmentHealthHistoryForms()
+    public function testCanSeeAllAppointments()
     {
         $baseUrl = 'test.clinect.com';
 
@@ -18,19 +18,19 @@ class HealthHistoryFormTests extends TestCase
 
         $connector->withMockClient($this->client($baseUrl));
 
-        $request = $connector->appointments('appointment-id')->withPracticeId('practice-id')->healthHistoryForms()->get();
+        $request = $connector->appointments()->get();
 
         $response = $connector->send($request);
 
         $this->assertSame($response->status(), 200);
 
         foreach ($response->json() as $key => $result) {
-            $this->assertSame($this->healthHistoryForms()[$key]['name'], $result['name']);
-            $this->assertSame($this->healthHistoryForms()[$key]['category'], $result['category']);
+            $this->assertSame($this->all()[$key]['name'], $result['name']);
+            $this->assertSame($this->all()[$key]['category'], $result['category']);
         }
     }
 
-    public function testCanSeeAppointmentHealthHistoryForm()
+    public function testCanSeeAppointment()
     {
         $baseUrl = 'test.clinect.com';
 
@@ -38,16 +38,16 @@ class HealthHistoryFormTests extends TestCase
 
         $connector->withMockClient($this->client($baseUrl));
 
-        $request = $connector->appointments('appointment-id')->withPracticeId('practice-id')->healthHistoryForms('id-3')->get();
+        $request = $connector->appointments('id-3')->get();
 
         $response = $connector->send($request);
 
         $this->assertSame($response->status(), 200);
-        $this->assertSame($response->json('name'), 'Appointment health history 3');
-        $this->assertSame($response->json('category'), 'appointment-health-history-3');
+        $this->assertSame($response->json('name'), 'Appointment 3');
+        $this->assertSame($response->json('category'), 'appointment-3');
     }
 
-    public function testAppointmentHealthHistoryFormNotFound()
+    public function testAppointmentNotFound()
     {
         $baseUrl = 'test.clinect.com';
 
@@ -55,7 +55,7 @@ class HealthHistoryFormTests extends TestCase
 
         $connector->withMockClient($this->client($baseUrl));
 
-        $request = $connector->appointments('appointment-id')->withPracticeId('practice-id')->healthHistoryForms('id-4')->get();
+        $request = $connector->appointments('id-4')->get();
 
         $response = $connector->send($request);
