--- tests/Feature/Resources/Requests/PersonTests.php
@@ -1,65 +1,0 @@
-<?php
-
-namespace Clinect\NextGen\Tests\Feature\Resources\Requests;
-
-use Clinect\NextGen\NextGen;
-use Orchestra\Testbench\TestCase;
-use Clinect\NextGen\Tests\Stubs\Person as PersonStub;
-
-class PersonTests extends TestCase
-{
-    use PersonStub;
-
-    public function testCanSeeAllPersons()
-    {
-        $baseUrl = 'test.clinect.com';
-
-        $connector = new NextGen(baseUrl: $baseUrl);
-
-        $connector->withMockClient($this->client($baseUrl));
-
-        $request = $connector->persons()->get();
-
-        $response = $connector->send($request);
-
-        $this->assertSame($response->status(), 200);
-
-        foreach ($response->json() as $key => $result) {
-            $this->assertSame($this->persons()[$key]['name'], $result['name']);
-            $this->assertSame($this->persons()[$key]['category'], $result['category']);
-        }
-    }
-
-    public function testCanSeePerson()
-    {
-        $baseUrl = 'test.clinect.com';
-
-        $connector = new NextGen(baseUrl: $baseUrl);
-
-        $connector->withMockClient($this->client($baseUrl));
-
-        $request = $connector->persons('id-3')->get();
-
-        $response = $connector->send($request);
-
-        $this->assertSame($response->status(), 200);
-        $this->assertSame($response->json('name'), 'Person 3');
-        $this->assertSame($response->json('category'), 'person-3');
-    }
-
-    public function testPersonNotFound()
-    {
-        $baseUrl = 'test.clinect.com';
-
-        $connector = new NextGen(baseUrl: $baseUrl);
-
-        $connector->withMockClient($this->client($baseUrl));
-
-        $request = $connector->persons('id-4')->get();
-
-        $response = $connector->send($request);
-
-        $this->assertSame($response->status(), 404);
-        $this->assertSame($response->json('error'), 'No data available');
-    }
-}
